Refactor Challenge in view models to Challenges
