+++ src/GRA.Controllers/ViewModel/Challenges/ChallengesDetailViewModel.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc.Rendering;
 using System.Collections.Generic;
 
-namespace GRA.Controllers.ViewModel.Challenge
+namespace GRA.Controllers.ViewModel.Challenges
 {
-    public class ChallengeDetailViewModel
+    public class ChallengesDetailViewModel
     {
         public GRA.Domain.Model.Challenge Challenge { get; set; }
         public GRA.Domain.Model.ChallengeTask Task { get; set; }
