+++ src/GRA.Web/Areas/MissionControl/Views/Challenges/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model GRA.Controllers.ViewModel.Challenge.ChallengeDetailViewModel
+﻿@model GRA.Controllers.ViewModel.Challenges.ChallengesDetailViewModel
 
 <form asp-controller="Challenges" method="post" role="form">
